push b4 work

+++ include/Simflow.hpp
@@ -318,6 +318,6 @@ void compute_fluid(const std::vector<Cell>& my_old_cells);
 
 void compute_fluid(float total_cells);
 
-void compute_fluid_time(float total_cells);
+void compute_fluid_time(long total_cells);
 
 #endif //SPEC_PROBLEM_HPP
